Dynamic Animations
We seem to want to cater this theme to the end user, so why not allow
animations to be chosen?

Included 11 presets to get started, but extremely easy to add/modify
them as needed!
I've also altered the settings.yml file to include the possibility to
set which animation to use.

I am running Jekyll 2.1.1 so my all.sass file has the settings defined
there, with the Jekyll paths commented out beside (sorry!)

Further, a typo was fixed in function.js that has been bugging me for a
while.

Enjoy!