(Almost) Final Commit
Everything is done in here EXCEPT for the unifying of the work
directory. The _includes folder does not compile into the site
directory.

There is a way to do so using config.yml (or apparently) but I cannot
get it to work correctly. Perhaps someone else can help out there?!

I have made the edit that I think should be made, but clearly something
is not working with it!