Updated _includes/clients.html
Although it’s absolutely not necessary, I like to make use of HTML
elements that are designed to do a specific job. In this case, one
could argue that the client’s photo is a figure, and the caption to
figure is who the client is and where he is from. Not only does this
help with Google and other search engines reading your site and
understanding your content, but it also focuses on accessibility. If
you can make it easier for a potential client who uses a screen reader
to read your content, I’m sure he or she would be more prone to hiring
you.