Added Google Form email subscription signup
Using the power of google forms, I have created a form that you can directly put in your site. It's actually pretty simple to do, so here are the instructions:

I dont have jekyll and I have allot of homework so I don't have time to set it up but this code should work pretty well. 
It probably* needs prettyfication, so I leave it up to you to do that! I just saw your latest vid and knew that I HAD to do this.
*definitely

It gives the responses to a google form, which is in my account but there is a link below to see responses. From there you can just 
cop and paste all the users. I would recommend users of this actually doing it themselves, though, it's not that hard.

How To Do:
1) create google form
2) click view live form
3) view source codes
4) copy everything inside the form tag
5) cut out all the crap
6) HAVEN'T GOTTEN HERE THIS PART IS TIME CONSUMING AND NEEDS JEKYLL OR SOMETHING make it pretty using normal techniques (sass, css, less, whatever)
7) enjoy!

Note that i haven't had a chance to make sure it works, but I am 90% sure it will. I am also 90% sure that it looks awful. As I said, I don't have that much time. If I get more time I may fix it later. We'll see.

~Albatross Games, AKA The Miner, AKA Russ
See responses: https://docs.google.com/spreadsheets/d/1PKD_eA15NRi_29M-tm-vpfL3mxIsc030O5JLPTLBNpQ/edit?usp=sharing